requirements.txt
streamlit
pandas
numpy
matplotlib
streamlit-authenticator
gcsfs
st-files-connection
oauth2client==4.1.3
google-api-python-client==2.64.0
plotly
openpyxl
statsmodels